feat(part4): 4.21
set middleware token check for delete

+++ part4/blog-list/controllers/blogs.js
@@ -12,9 +12,10 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body;
-    const decodedToken = jwt.verify(request.token , process.env.SECRET)
+    const decodedToken = await jwt.verify(request.token , process.env.SECRET)
 
-    if(!token || !decodedToken.id) {
+    console.log(request.token)
+    if(!decodedToken.id) {
         return response.status(401).json({error: 'token missing or invalid'})
     }
 
@@ -45,9 +46,17 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
     let idToDelete = request.params.id;
-    const result = await Blog.findByIdAndRemove(idToDelete);
-    response.status(204).end();
-
+    const decodedToken = jwt.verify(request.token , process.env.SECRET)
+    
+    const blogToCheckBeforeRemove = await Blog.findById(idToDelete)
+    const user = await User.findById(decodedToken.id)
+    
+    if(blogToCheckBeforeRemove.user.toString() === user._id.toString()) {
+        const result = await Blog.findByIdAndRemove(idToDelete);
+        response.status(204).end();
+    }
+    
+    response.status(401).json({error: "Unauthorizated user for this operation!"}).end();
 });
 
 blogsRouter.put('/:id', async (request,response) => {
